backend/pyproject.toml
[tool.poetry]
name = "inspirehep"
version = "0.1.19"
description = "NSPIRE is the leading information platform for High Energy Physics (HEP) literature which provides users with high quality, curated content covering the entire corpus of HEP literature and the fulltext of all such articles that are Open Access."
authors = ["CERN <admin@inspirehep.net>"]
license = "GPL-2.0"

[tool.poetry.dependencies]
celery = "<=4.2.1"
# inspirehep = {path = ".",editable = true}
# Click in version 7 have broken progressbar in python 3
# waiting for this PR to be merged https://github.com/pallets/click/pull/1163
click = "<7.0.0"
python = ">=3.6,<3.7"
invenio = {version = "~=3.1.0,==3.1.0", extras = [ "auth"]}
invenio-db = {version = "*", extras = [ "postgresql", "versioning"]}
invenio-search = {version = "*", extras = [ "elasticsearch5",]}
invenio-app = "~=1.1.0,>=1.1.0"
invenio-access = "~=1.1.0,>=1.1.0"
invenio-rest = ">=1.0.0,<1.1.0"
invenio-indexer = ">=1.0.1,<1.0.2"
invenio-jsonschemas = ">=1.0.0,<1.1.0"
invenio-pidstore = ">=1.0.0,<1.1.0"
invenio-records-rest = ">=1.2.1,<1.3.0"
timeexecution = "<4.0"
invenio-records = ">=1.0.0,<1.1.0"
invenio-records-files = ">=1.0.0a10"
inspire-dojson = "~=62.0.0,>=62.0.0"
inspire-schemas = "~=60.0"
inspire-utils = "~=3.0.0,>=3.0.5"
inspire-query-parser = "~=6.0.0"
uwsgi = ">=2.0"
marshmallow = "<3.0.0"
flask-shell-ipython = "~=0.3,>=0.3.0"
flask-sqlalchemy = {git = "https://github.com/inspirehep/flask-sqlalchemy.git"}
orcid = "==1.0.3"
inspire-service-orcid = {git = "https://github.com/drjova/inspire-service-orcid.git"}
node-semver = "<0.2.0,>=0.1.1"
pybtex = "*"
cryptography = "<2.6"
# msgpack = "*"  #  FIXME:  We want to use it but it might mess with invenio-indexer which is forcing msgpack-python https://github.com/inveniosoftware/invenio-celery/issues/58
flask-kvsession = {git = "https://github.com/inspirehep/flask-kvsession.git"}
hepcrawl = "^13.0"
python-redis-lock = "^3.3"
invenio-base = "^1.0"
inspire-json-merger = "^11.0"
lxml = "^4.3"
fqn-decorators = "~=1.2,>=1.2.3"
editdistance = "^0.5.3"
munkres = "^1.0"
rt = "^1.0"
gunicorn = "^19.0.0"
sentry-sdk = {version = "^0.10.2",extras = ["flask"]}
prometheus-flask-exporter = "^0.9.0"

[tool.poetry.dev-dependencies]
freezegun = "*"
coverage = ">=4.4.1"
isort = ">=4.3"
mock = ">=2.0.0"
pydocstyle = ">=2.0.0"
pytest = ">=3.3.1,<4.0.0"
pytest-cov = "==2.6.1"
pytest-invenio = ">=1.1.0,<1.2.0"
pytest-mock = ">=1.6.0"
pytest-random-order = ">=0.5.4"
pytest-datadir = "*"
pytest-runner = ">=3.0.0,<5"
remote-pdb = "*"
black = "==18.9b0"
factory_boy = "~=2.11.1,>=2.11.1"
requests-mock = "~=1.5.2,>=1.5.2"
pytest-vcr = "*"
pre-commit = "*"
pylint = "*"
flake8 = "*"
watchdog = "*"
virtualenv = "<16.5.0"
ipdb = "^0.12.0"

[build-system]
requires = ["poetry==0.12.16"]
build-backend = "poetry.masonry.api"
